+++ src/Ui/Bundle/EventBundle/Resources/assets/js/components/agenda/_AgendaAllSheet.js
@@ -1,0 +1,14 @@
+var Meet = require('./_Meet');
+
+function AgendaAllSheet(element) {
+    this.element = element;
+    this.meetingModal = document.getElementById('meeting-modal');
+    this.addMeet = this.addMeet.bind(this);
+    Array.prototype.forEach.call(this.element.querySelectorAll('.meet'), this.addMeet);
+}
+
+AgendaAllSheet.prototype.addMeet = function (element) {
+    new Meet(null, element, this.meetingModal);
+};
+
+module.exports = AgendaAllSheet;